+++ src/app/admin-module/admin.module.ts
@@ -8,8 +8,9 @@ import { AutocompleteModule } from '../component-modules/autocomplete/autocomple
 import { BreadcrumbComponent } from './components/breadcrumb.component';
 import { DashboardComponent } from './components/dashboard.component';
 import { RecipeDetailCheckboxListComponent } from './components/recipe-detail-checkbox-list.component';
+import { RecipeIngredientDialogComponent } from './components/recipe-ingredient-dialog/recipe-ingredient-dialog.component';
 import { RecipeListComponent } from './components/recipe-list.component';
-import { RecipeDetailComponent } from './components/recipe-detail.component';
+import { RecipeDetailComponent } from './components/recipe-detail/recipe-detail.component';
 import { DataCollectionModule } from '../data-collection/data-collection.module';
 import { RecipeTitleDialogComponent } from './components/recipe-title-dialog.component';
 import { SourceDetailComponent } from './components/source-detail.component';
@@ -31,6 +32,7 @@ import { SharedMaterialModule } from '../shared/shared-material.module';
     BreadcrumbComponent,
     RecipeTitleDialogComponent,
     RecipeDetailCheckboxListComponent,
+    RecipeIngredientDialogComponent,
   ],
   imports: [
     ReactiveFormsModule,