+++ src/app/admin-module/models/data-collection-page-data.interface.ts
@@ -1,5 +1,4 @@
 export interface DataCollectionPageData {
   title: string;
   ingredients: any[]; // todo
-
 }