+++ src/app/data-collection/components/data-field/data-field.component.ts
@@ -1,5 +1,8 @@
 import { Component, Input, ViewEncapsulation } from '@angular/core';
 import { FormGroup } from '@angular/forms';
+import { DataFieldCustomAction } from '../../../admin-module/models/data-field-custom-action.type';
+import { LanguageService } from '../../../shared/services/language-service/language.service';
+import locale from './data-field.locale.json';
 
 @Component({
   selector: 'app-data-field',
@@ -33,7 +36,7 @@ import { FormGroup } from '@angular/forms';
           class="px-2 rounded bg-blue-700 text-white text-sm font-medium"
           (click)="toNull()"
         >
-          Set to NULL
+          {{ locale[lang.language].NullButtonLabel }}
         </button>
       </div>
     </div>
@@ -41,13 +44,18 @@ import { FormGroup } from '@angular/forms';
   encapsulation: ViewEncapsulation.None,
 })
 export class DataFieldComponent {
+  // TODO (nice to have): what if pass here FormControl instead FormGroup and control key?
   /** Form reference */
   @Input({ required: true }) formGroupRef!: FormGroup;
   @Input({ required: true }) formControlNameRef!: string;
   /** Label for input field */
   @Input({ required: true }) title!: string;
-  //
-  @Input() customActionList: any[] | undefined; // TODO type - label, click handler
+  /** List of custom actions */
+  @Input() customActionList: DataFieldCustomAction[] | undefined;
+
+  protected readonly locale = locale;
+
+  constructor(protected readonly lang: LanguageService) {}
 
   /** Set form control value to NULL */
   protected toNull() {